refactor: 테스트 위치 변경 및 DisplayName 설명 상세화
