+++ src/test/java/racingcar/domain/Car/CarNameTest.java
@@ -1,5 +1,6 @@
 package racingcar.domain.Car;
 
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -9,6 +10,7 @@ import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.stream.Stream;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -20,6 +22,23 @@ class CarNameTest {
     void carNameNullOrEmptyTest(String input) {
         assertThatThrownBy(() -> new CarName(input))
                 .isInstanceOf(RuntimeException.class);
+    }
+
+    @ParameterizedTest
+    @DisplayName("올바른 이름으로 CarName 생성 성공 테스트")
+    @ValueSource(strings = {"bepoz  ", "12345", " joy", "b ank"})
+    void generate_validName(String input) {
+        Car car = new Car(input);
+        assertThat(car.getCarName().getName()).isEqualTo(input.trim());
+    }
+
+    @ParameterizedTest
+    @DisplayName("이름이 6글자 이상일 때의 CarName 생성 테스트")
+    @ValueSource(strings = {"123456"})
+    void generate_invalidName(String input) {
+        Assertions.assertThatThrownBy(() -> {
+            new Car(input);
+        }).isInstanceOf(RuntimeException.class);
     }
 
     private static Stream<Arguments> provideEmptyAndBlankAndNullTestCase() {
